+++ optd-datafusion-repr/src/lib.rs
@@ -87,6 +87,7 @@ impl DatafusionOptimizer {
             Arc::new(EliminateLimitRule::new()),
             Arc::new(EliminateDuplicatedSortExprRule::new()),
             Arc::new(EliminateDuplicatedAggExprRule::new()),
+            Arc::new(ProjectMergeRule::new()),
         ]
     }
 
@@ -97,7 +98,7 @@ impl DatafusionOptimizer {
         for rule in rules {
             rule_wrappers.push(RuleWrapper::new_cascades(rule));
         }
-        // rule_wrappers.push(RuleWrapper::new_cascades(Arc::new(
+        // rule_wrappers.push(RuleWrapper::new_heuristic(Arc::new(
         //     ProjectMergeRule::new(),
         // )));
         rule_wrappers.push(RuleWrapper::new_cascades(Arc::new(
