+++ optd-datafusion-repr/src/rules.rs
@@ -1,5 +1,4 @@
 // mod filter_join;
-mod projection_pushdown;
 mod eliminate_duplicated_expr;
 mod eliminate_limit;
 mod filter;
@@ -7,10 +6,14 @@ mod filter_pushdown;
 mod joins;
 mod macros;
 mod physical;
+mod project_transpose;
 
 // pub use filter_join::FilterJoinPullUpRule;
-pub use projection_pushdown::{
-    ProjectMergeRule, ProjectFilterTransposeRule
+pub use project_transpose::{
+    project_merge::ProjectMergeRule, 
+    project_filter_transpose::ProjectFilterTransposeRule,
+    project_join_transpose::ProjectionPullUpJoin,
+    filter_project_transpose::FilterProjectTransposeRule,
 };
 pub use eliminate_duplicated_expr::{
     EliminateDuplicatedAggExprRule, EliminateDuplicatedSortExprRule,
@@ -19,9 +22,9 @@ pub use eliminate_limit::EliminateLimitRule;
 pub use filter::{EliminateFilterRule, SimplifyFilterRule, SimplifyJoinCondRule};
 pub use filter_pushdown::{
     FilterAggTransposeRule, FilterCrossJoinTransposeRule, FilterInnerJoinTransposeRule,
-    FilterMergeRule, FilterProjectTransposeRule, FilterSortTransposeRule,
+    FilterMergeRule, FilterSortTransposeRule,
 };
 pub use joins::{
-    EliminateJoinRule, HashJoinRule, JoinAssocRule, JoinCommuteRule, ProjectionPullUpJoin,
+    EliminateJoinRule, HashJoinRule, JoinAssocRule, JoinCommuteRule,
 };
 pub use physical::PhysicalConversionRule;
